Fixed generation of error log and include

+++ asseterrorlog_generator.rb
@@ -4,15 +4,19 @@ module Jekyll
     safe true
     
     def generate(site)
-      error_log = File.expand_path(File.join(site.config["source"],"_includes","asset_pipeline_errors"))
-      File.open(error_log, 'w') do |f|
-        f.write <<END
+      include_dir = File.expand_path(File.join(site.config["source"],"_includes"))
+      if File.exists?(include_dir)
+	      error_log = File.join(include_dir, "asset_pipeline_errors")
+	      # Write the file unless it already exists
+	      File.open(error_log, 'w') do |f|
+	        f.write <<END
 {% capture asset_pipeline_error_content %}{% include asset_pipeline_errors.log %}{% endcapture %}
 {% unless asset_pipeline_error_content contains "Liquid error" or asset_pipeline_error_content contains "not found in _includes directory" %}
-{% include asset_pipeline_errors.log %}</span>
+{% include asset_pipeline_errors.log %}
 {% endunless %}
 END
-      end unless File.exists?(error_log)
+	      end unless File.exists?(error_log)
+	   end
     end
 
   end
